# services/api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "این سرویس به عنوان درگاه اصلی برای دسترسی به میکروسرویس‌های امانت و کاربران عمل می‌کند.",
        "title": "Books \u0026 Trust API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/loans": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new loan record by calling the internal loan microservice. Only the user themselves can initiate the loan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Create a new loan",
                "parameters": [
                    {
                        "description": "Create Loan Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http_loan_handler.CreateLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "loan successfully created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/contracts.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data_message": {
                                            "$ref": "#/definitions/internal_handler_http_loan_handler.CreateLoanResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "invalid create loan request",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "403": {
                        "description": "you can not create loan",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a loan. Only the Lender (UserID) or Borrower (OwnerID) can update it. Note: The borrower cannot extend the deadline.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Update an existing loan",
                "parameters": [
                    {
                        "description": "Update Loan Request Body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler_http_loan_handler.UpdateLoanRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "loan updated successfully",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "invalid update loan request",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "403": {
                        "description": "you can not update this loan / borrower cannot extend the deadline",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "404": {
                        "description": "loan record not found",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    }
                }
            }
        },
        "/loans/owner/{ownerID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all loans shared or owned by a specific user. Users can only fetch their own loan lists (ownerID must match the authenticated user ID).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "List loans by owner ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Owner User UUID",
                        "name": "ownerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "loans fetched successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/contracts.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data_message": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/internal_handler_http_loan_handler.Loan"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "invalid url param",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "403": {
                        "description": "you cannot access these loans",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    }
                }
            }
        },
        "/loans/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Fetches the details of a specific loan using its UUID from the URL parameter. Only the Lender (UserID) or Borrower (OwnerID) involved in the loan can access it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Get loan details by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "loan fetched successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/contracts.ResponseSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data_message": {
                                            "$ref": "#/definitions/internal_handler_http_loan_handler.GetLoanByIDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "invalid url param",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "403": {
                        "description": "you can not get loan",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "404": {
                        "description": "loan record not found",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    }
                }
            }
        },
        "/loans/{id}/claim": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an authenticated user to claim/borrow an available book loan by its ID. The user cannot be the owner of the loan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loans"
                ],
                "summary": "Claim an available loan position",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Loan UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user banned successfully",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseSuccess"
                        }
                    },
                    "400": {
                        "description": "invalid url param",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "403": {
                        "description": "forbidden - cannot claim own loan or user is banned",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "404": {
                        "description": "loan record not found",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "409": {
                        "description": "conflict - loan already claimed by another user",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#/definitions/contracts.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "contracts.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "contracts.ResponseSuccess": {
            "type": "object",
            "properties": {
                "data_message": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_http_loan_handler.CreateLoanRequest": {
            "type": "object",
            "required": [
                "book_name",
                "owner_id"
            ],
            "properties": {
                "book_name": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                }
            }
        },
        "internal_handler_http_loan_handler.CreateLoanResponse": {
            "type": "object",
            "required": [
                "book_name",
                "id",
                "user_id"
            ],
            "properties": {
                "book_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "owner_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_handler_http_loan_handler.GetLoanByIDResponse": {
            "type": "object",
            "required": [
                "loan"
            ],
            "properties": {
                "loan": {
                    "$ref": "#/definitions/internal_handler_http_loan_handler.Loan"
                }
            }
        },
        "internal_handler_http_loan_handler.Loan": {
            "type": "object",
            "required": [
                "book_name",
                "id",
                "user_id"
            ],
            "properties": {
                "book_name": {
                    "type": "string",
                    "example": "The Go Programming Language"
                },
                "created_at": {
                    "type": "string",
                    "example": "2026-06-03T12:00:00Z"
                },
                "deadline": {
                    "type": "string",
                    "example": "2026-06-16T18:30:00Z"
                },
                "id": {
                    "type": "string",
                    "example": "loan-uuid-111"
                },
                "owner_id": {
                    "type": "string",
                    "example": "owner-uuid-456"
                },
                "status": {
                    "type": "string",
                    "example": "BORROWED"
                },
                "user_id": {
                    "type": "string",
                    "example": "user-uuid-123"
                }
            }
        },
        "internal_handler_http_loan_handler.UpdateLoanRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "book_name": {
                    "type": "string"
                },
                "deadline": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "تایپ کنید: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}